Store downloads in the database not the file system - refs #2964

+++ wp-content/themes/bp-child/functions/tmp_functions.php
@@ -229,12 +229,14 @@ if(is_super_admin())
         {
             $downloads = $wpdb->get_results("SELECT * FROM " . $wpdb->prefix . "bp_groups_downloads");
             foreach ($downloads as $row) {
-                $data = array();
-                $data['download_file'] = file_get_contents($row->location);
-                $data = stripslashes_deep( $data );
-                
-                $wpdb->update($wpdb->prefix . 'bp_groups_downloads', $data, array('id' => $row->id));
-                unlink($row->location);
+                if (!$row->download_file) {
+                    $data = array();
+                    $data['download_file'] = file_get_contents($row->location);
+                    $data = stripslashes_deep( $data );
+                    
+                    $wpdb->update($wpdb->prefix . 'bp_groups_downloads', $data, array('id' => $row->id));
+                    unlink($row->location);
+                }
             }
         }        
     }
